Refactor controllers and pass accepted file types
