+++ app-modules/ai/src/Http/Controllers/AssistantWidget/AuthorizeBroadcastController.php
@@ -17,7 +17,7 @@
       in the software, and you may not remove or obscure any functionality in the
       software that is protected by the license key.
     - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      of the licensor in the software. Any use of the licensor's trademarks is subject
       to applicable law.
     - Canyon GBS Inc. respects the intellectual property rights of others and expects the
       same in return. Canyon GBS® and Aiding App® are registered trademarks of
@@ -45,9 +45,9 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Str;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
-class AssistantBroadcastController extends Controller
+class AuthorizeBroadcastController extends Controller
 {
-    public function auth(Request $request, Broadcaster $broadcaster): mixed
+    public function __invoke(Request $request, Broadcaster $broadcaster): mixed
     {
         if ($request->hasSession()) {
             $request->session()->reflash();
