+++ app-modules/ai/src/Http/Controllers/AssistantWidget/GetWidgetConfigController.php
@@ -1,0 +1,85 @@
+<?php
+
+/*
+<COPYRIGHT>
+
+    Copyright © 2016-2026, Canyon GBS Inc. All rights reserved.
+
+    Aiding App® is licensed under the Elastic License 2.0. For more details,
+    see <https://github.com/canyongbs/aidingapp/blob/main/LICENSE.>
+
+    Notice:
+
+    - You may not provide the software to third parties as a hosted or managed
+      service, where the service provides users with access to any substantial set of
+      the features or functionality of the software.
+    - You may not move, change, disable, or circumvent the license key functionality
+      in the software, and you may not remove or obscure any functionality in the
+      software that is protected by the license key.
+    - You may not alter, remove, or obscure any licensing, copyright, or other notices
+      of the licensor in the software. Any use of the licensor's trademarks is subject
+      to applicable law.
+    - Canyon GBS Inc. respects the intellectual property rights of others and expects the
+      same in return. Canyon GBS® and Aiding App® are registered trademarks of
+      Canyon GBS Inc., and we are committed to enforcing and protecting our trademarks
+      vigorously.
+    - The software solution, including services, infrastructure, and code, is offered as a
+      Software as a Service (SaaS) by Canyon GBS Inc.
+    - Use of this software implies agreement to the license terms and conditions as stated
+      in the Elastic License 2.0.
+
+    For more information or inquiries please visit our website at
+    <https://www.canyongbs.com> or contact us via email at legal@canyongbs.com.
+
+</COPYRIGHT>
+*/
+
+namespace AidingApp\Ai\Http\Controllers\AssistantWidget;
+
+use AidingApp\Portal\Settings\PortalSettings;
+use App\Http\Controllers\Controller;
+use Filament\Support\Colors\Color;
+use Illuminate\Http\JsonResponse;
+use Illuminate\Http\Request;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\URL;
+
+class GetWidgetConfigController extends Controller
+{
+    public function __invoke(Request $request): JsonResponse
+    {
+        $manifestPath = public_path('storage/widgets/assistant/.vite/manifest.json');
+        /** @var array<string, array{file: string, name: string, src: string, isEntry: bool}> $manifest */
+        $manifest = json_decode(File::get($manifestPath), true, 512, JSON_THROW_ON_ERROR);
+
+        $widgetEntry = $manifest['src/widget.js'];
+
+        $settings = app(PortalSettings::class);
+
+        $websocketsConfig = config('filament.broadcasting.echo');
+
+        return response()->json([
+            'asset_url' => route('widgets.assistant.asset'),
+            'js' => route('widgets.assistant.asset', ['file' => $widgetEntry['file']]),
+            'send_message_url' => route('widgets.assistant.api.messages'),
+            'websockets_config' => $websocketsConfig,
+            'auth_endpoint' => route('widgets.assistant.api.broadcasting.auth'),
+            'primary_color' => collect(Color::all()[$settings->knowledge_management_portal_primary_color->value ?? 'blue'])
+                ->map(Color::convertToRgb(...))
+                ->map(fn (string $value): string => (string) str($value)->after('rgb(')->before(')'))
+                ->all(),
+            'rounding' => $settings->knowledge_management_portal_rounding->value ?? 'md',
+            'is_authenticated' => (bool) auth('contact')->user(),
+            'authenticate_request_url' => auth('contact')->user() ? null : URL::to(
+                URL::signedRoute(
+                    name: 'widgets.assistant.api.authenticate.request',
+                    absolute: false,
+                )
+            ),
+            'service_request_types_url' => $settings->knowledge_management_portal_service_management
+                ? route('widgets.assistant.api.service-request-types')
+                : null,
+            'portal_service_management' => (bool) $settings->knowledge_management_portal_service_management,
+        ]);
+    }
+}
