+++ app-modules/ai/src/Http/Controllers/AssistantWidget/SendMessageController.php
@@ -17,7 +17,7 @@
       in the software, and you may not remove or obscure any functionality in the
       software that is protected by the license key.
     - You may not alter, remove, or obscure any licensing, copyright, or other notices
-      of the licensor in the software. Any use of the licensor’s trademarks is subject
+      of the licensor in the software. Any use of the licensor's trademarks is subject
       to applicable law.
     - Canyon GBS Inc. respects the intellectual property rights of others and expects the
       same in return. Canyon GBS® and Aiding App® are registered trademarks of
@@ -38,82 +38,15 @@ namespace AidingApp\Ai\Http\Controllers\AssistantWidget;
 
 use AidingApp\Ai\Jobs\PortalAssistant\SendMessage;
 use AidingApp\Ai\Models\PortalAssistantThread;
-use AidingApp\Portal\Settings\PortalSettings;
 use App\Http\Controllers\Controller;
-use Filament\Support\Colors\Color;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
-use Illuminate\Support\Facades\File;
-use Illuminate\Support\Facades\Storage;
-use Illuminate\Support\Facades\URL;
 use Illuminate\Support\Str;
-use Symfony\Component\HttpFoundation\StreamedResponse;
 
-class AssistantWidgetController extends Controller
+class SendMessageController extends Controller
 {
-    public function config(Request $request): JsonResponse
-    {
-        $manifestPath = public_path('storage/widgets/assistant/.vite/manifest.json');
-        /** @var array<string, array{file: string, name: string, src: string, isEntry: bool}> $manifest */
-        $manifest = json_decode(File::get($manifestPath), true, 512, JSON_THROW_ON_ERROR);
-
-        $widgetEntry = $manifest['src/widget.js'];
-
-        $settings = app(PortalSettings::class);
-
-        $websocketsConfig = config('filament.broadcasting.echo');
-
-        return response()->json([
-            'asset_url' => route('widgets.assistant.asset'),
-            'js' => route('widgets.assistant.asset', ['file' => $widgetEntry['file']]),
-            'send_message_url' => route('widgets.assistant.api.messages'),
-            'websockets_config' => $websocketsConfig,
-            'auth_endpoint' => route('widgets.assistant.api.broadcasting.auth'),
-            'primary_color' => collect(Color::all()[$settings->knowledge_management_portal_primary_color->value ?? 'blue'])
-                ->map(Color::convertToRgb(...))
-                ->map(fn (string $value): string => (string) str($value)->after('rgb(')->before(')'))
-                ->all(),
-            'rounding' => $settings->knowledge_management_portal_rounding->value ?? 'md',
-            'is_authenticated' => (bool) auth('contact')->user(),
-            'authenticate_request_url' => auth('contact')->user() ? null : URL::to(
-                URL::signedRoute(
-                    name: 'widgets.assistant.api.authenticate.request',
-                    absolute: false,
-                )
-            ),
-            'service_request_types_url' => $settings->knowledge_management_portal_service_management
-                ? route('widgets.assistant.api.service-request-types')
-                : null,
-            'portal_service_management' => (bool) $settings->knowledge_management_portal_service_management,
-        ]);
-    }
-
-    public function asset(Request $request, string $file): StreamedResponse
-    {
-        $path = "widgets/assistant/{$file}";
-
-        $disk = Storage::disk('public');
-
-        abort_if(! $disk->exists($path), 404, 'File not found.');
-
-        $mimeType = $disk->mimeType($path);
-
-        $stream = $disk->readStream($path);
-
-        abort_if(is_null($stream), 404, 'File not found.');
-
-        return response()->streamDownload(
-            function () use ($stream) {
-                fpassthru($stream);
-                fclose($stream);
-            },
-            $file,
-            ['Content-Type' => $mimeType]
-        );
-    }
-
-    public function sendMessage(Request $request): JsonResponse
+    public function __invoke(Request $request): JsonResponse
     {
         $data = $request->validate([
             'content' => ['required', 'string', 'max:25000'],
